+++ src/Libraries/SmartStore.Services/Filter/FilterCriteria.cs
@@ -38,6 +38,7 @@ namespace SmartStore.Services.Filter
 
 		// Metadata
 		public int MatchCount { get; set; }
+		public int DisplayOrder { get; set; }
 		public bool IsInactive { get; set; }
 		public string NameLocalized { get; set; }
 		public string ValueLocalized { get; set; }